refactor(ui): split ApiDetailView into per-section files
The 835-line monolith mixed ten concerns: cross-section data
fetching, the operations state machine (filter / tag / pagination /
all-ops prefetch), six section renderers, an OpenAPI security-scheme
formatter, and two skeletons. Pull each into its own file under
`components/workspace/api-detail/` so the orchestrator becomes ~190
lines that reads top-to-bottom: queries, then composition.

The operations widget is the only sub-component that genuinely owns
state — everything else is presentational. Extracted that state
machine into a `useApiOperations` hook so the orchestrator can still
surface the totals (`opsTotal`, `tagOptions.length`) to peer
components (the overview strip and `ApiSummary`) without a
side-effecty parent-callback bridge.

No behaviour change. All 443 unit tests + 33 e2e specs still pass.

- New folder `components/workspace/api-detail/`:
  - `ApiDetailView.tsx` — orchestrator (queries + composition)
  - `OverviewStrip.tsx` + `MetaItem.tsx` — top stats row
  - `CredentialsSection.tsx`
  - `ToolkitsSection.tsx` (+ `BoundToolkit` type)
  - `OperationsSection.tsx` + `OperationRow.tsx` + `useApiOperations.ts`
  - `WorkflowsSection.tsx`
  - `SecuritySchemesSection.tsx` (with `formatSchemeType` helper)
  - `skeletons.tsx` (`ApiDetailSkeleton`, `OperationsSkeleton`)
  - `index.ts` — re-exports `ApiDetailView`
- Delete the old `components/workspace/ApiDetailView.tsx` monolith
- Update `pages/ApiDetailPage.tsx` to import from the new barrel
