+++ ui/src/components/workspace/api-detail/index.ts
@@ -1,0 +1,1 @@
+export { ApiDetailView } from './ApiDetailView';
